refactor(run): extract repl submodule (Wave C6)
Moves the interactive agent REPL loop (run_agent_repl, ~268 lines) into
run/repl.rs. run.rs 2,654 -> 2,385. Single contiguous function; no tests
reference it. Done out of plan order (before C5) because it's a clean
single-region cut, whereas C5/forge is a 7-region interleaved move best done
fresh -- and C1-C4 already turned the clusters run_agent_repl orchestrates into
clean re-exported submodules, which was the plan's stated precondition for
doing the REPL last.

Pure code movement (no A/B). 'use super::*' reaches run.rs's items the loop
drives (build_runtime_streaming, run_turn_with_retry, index_turn_for_recall,
maybe_compact_session, format_ctx_gauge, CliPrompter, EMPTY_RESPONSE_NUDGE,
save_session*, current_model, ...); explicit imports for std::io + crate::theme
+ crate::commands slash-dispatch. Re-export: pub use repl::run_agent_repl (lib.rs
+ main.rs entry point). run.rs's now-unused commands import dropped.

whole-crate build green; fmt + clippy (lean + --all-features) -D warnings clean;
cargo test --lib 1035 passed (unchanged). Stacked on #148 (C4).
